requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
boto==2.49.0
boto3==1.15.9
botocore==1.18.9
Django==3.0.8
django-crispy-forms==1.9.1
django-ipware==3.0.0
django-storages==1.10.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
future==0.18.2
gunicorn==20.0.4
jmespath==0.10.0
kombu==4.6.11
Pillow==7.2.0
python-dateutil==2.8.1
pytz==2020.1
reportlab==3.5.49
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
vine==1.3.0
whitenoise==5.2.0
